feat(models): Adds first Post models
