+++ PIUApp/PIUApp/Models/About.cs
@@ -1,0 +1,7 @@
+namespace PIUApp.Models;
+
+// Root myDeserializedClass = JsonConvert.DeserializeObject<List<Root>>(myJsonResponse);
+public class About
+{
+    public string href { get; set; }
+}
