+++ PIUApp/PIUApp/Models/Author.cs
@@ -1,0 +1,7 @@
+namespace PIUApp.Models;
+
+public class Author
+{
+    public bool embeddable { get; set; }
+    public string href { get; set; }
+}
