+++ PIUApp/PIUApp/Models/Collection.cs
@@ -1,0 +1,5 @@
+namespace PIUApp.Models;
+public class Collection
+{
+    public string href { get; set; }
+}
