+++ PIUApp/PIUApp/Models/Content.cs
@@ -1,0 +1,6 @@
+namespace PIUApp.Models;
+public class Content
+{
+    public string rendered { get; set; }
+    public bool @protected { get; set; }
+}
