+++ PIUApp/PIUApp/Models/Cury.cs
@@ -1,0 +1,7 @@
+namespace PIUApp.Models;
+public class Cury
+{
+    public string name { get; set; }
+    public string href { get; set; }
+    public bool templated { get; set; }
+}
