+++ PIUApp/PIUApp/Models/Excerpt.cs
@@ -1,0 +1,6 @@
+namespace PIUApp.Models;
+public class Excerpt
+{
+    public string rendered { get; set; }
+    public bool @protected { get; set; }
+}
