+++ PIUApp/PIUApp/Models/Guid.cs
@@ -1,0 +1,5 @@
+namespace PIUApp.Models;
+public class Guid
+{
+    public string rendered { get; set; }
+}
