+++ PIUApp/PIUApp/Models/PredecessorVersion.cs
@@ -1,0 +1,5 @@
+public class PredecessorVersion
+{
+    public int id { get; set; }
+    public string href { get; set; }
+}
