+++ PIUApp/PIUApp/Models/Reply.cs
@@ -1,0 +1,5 @@
+public class Reply
+{
+    public bool embeddable { get; set; }
+    public string href { get; set; }
+}
