+++ PIUApp/PIUApp/Models/Self.cs
@@ -1,0 +1,4 @@
+public class Self
+{
+    public string href { get; set; }
+}
