+++ PIUApp/PIUApp/Models/Title.cs
@@ -1,0 +1,4 @@
+public class Title
+{
+    public string rendered { get; set; }
+}
