+++ PIUApp/PIUApp/Models/VersionHistory.cs
@@ -1,0 +1,5 @@
+public class VersionHistory
+{
+    public int count { get; set; }
+    public string href { get; set; }
+}
