+++ PIUApp/PIUApp/Models/WpAttachment.cs
@@ -1,0 +1,4 @@
+public class WpAttachment
+{
+    public string href { get; set; }
+}
