+++ PIUApp/PIUApp/Models/WpFeaturedmedium.cs
@@ -1,0 +1,5 @@
+public class WpFeaturedmedium
+{
+    public bool embeddable { get; set; }
+    public string href { get; set; }
+}
