+++ PIUApp/PIUApp/Models/WpTerm.cs
@@ -1,0 +1,6 @@
+public class WpTerm
+{
+    public string taxonomy { get; set; }
+    public bool embeddable { get; set; }
+    public string href { get; set; }
+}
